% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_contours.R
\name{add_contours}
\alias{add_contours}
\title{Add Contours to Base R Plot}
\usage{
add_contours(he, params)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{params}{List}
}
\value{
plot side effect
}
\description{
Add Contours to Base R Plot
}
\keyword{aplot}
\keyword{internal}
