% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCEAweb.R
\name{BCEAweb}
\alias{BCEAweb}
\title{BCEAweb}
\usage{
BCEAweb(e = NULL, c = NULL, parameters = NULL, ...)
}
\arguments{
\item{e}{A matrix containing the simulations for the effectiveness
variable (with number of simulation rows and number of interventions
columns). Defaults at NULL, which means the user has to load their 
own values using the web-interface}

\item{c}{A matrix containing the simulations for the cost
variable (with number of simulation rows and number of interventions
columns). Defaults at NULL, which means the user has to load their 
own values using the web-interface}

\item{parameters}{A matrix with the simulations for all the relevant
model parameters. Defaults at NULL, which means the user has to load 
their own values using the web-interface}

\item{...}{Additional parameters.}
}
\description{
Launches the web-app
}
\examples{
\donttest{
data(Vaccine)
BCEAweb(e,c,vaccine)
}
}
\references{
Baio, G., Dawid, A. P. (2011). Probabilistic Sensitivity
Analysis in Health Economics.  Statistical Methods in Medical Research
doi:10.1177/0962280211419832.

Baio G. (2012). Bayesian Methods in Health Economics. CRC/Chapman Hall,
London
}
\seealso{
\code{\link{bcea}}
}
\author{
Gianluca Baio
}
\concept{Health evaluation}
\concept{Value of information}
