% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_interv_given_k.R
\name{best_interv_given_k}
\alias{best_interv_given_k}
\title{Select best option for each value of willingness to pay}
\usage{
best_interv_given_k(eib, ref, comp)
}
\arguments{
\item{eib}{Expected incremental benefit}
}
\value{

}
\description{
Select best option for each value of willingness to pay
}
\examples{

}
