% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce_table.R
\name{ce_table}
\alias{ce_table}
\title{Cost-effectiveness summary statistics table}
\usage{
ce_table(he, wtp = 25000, ...)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{wtp}{Willingness to pay}

\item{...}{Additional parameters}
}
\description{
As is commonly shown in a journal paper
}
\keyword{internal}
