% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceac.plot.R
\name{ceac.plot}
\alias{ceac.plot}
\title{Cost-Effectiveness Acceptability Curve (CEAC) Plot Generic}
\usage{
ceac.plot(he, ...)
}
\description{
Produces a plot of the Cost-Effectiveness Acceptability Curve (CEAC) against
the willingness to pay threshold.
}
\seealso{
\code{\link{ceac.plot.bcea}}
}
