% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceac_plot_graph.R
\name{ceac_matplot}
\alias{ceac_matplot}
\title{CEAC Matrix Plot}
\usage{
ceac_matplot(he, pos_legend, graph_params, ceac)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{pos_legend}{Legend position}

\item{graph_params}{Aesthetic ggplot parameters}

\item{ceac}{\code{ceac} index in \code{he}}
}
\description{
CEAC plot using \code{matplot} in Base R.
}
\keyword{hplot}
\keyword{internal}
