% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceef.summary.R
\name{ceef.summary}
\alias{ceef.summary}
\title{Summary table for CEEF}
\usage{
ceef.summary(he, frontier_data, frontier_params, include.ICER = FALSE, ...)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{frontier_data}{Frontier data}

\item{frontier_params}{Frontier parameters}

\item{include.ICER}{Should we include the ICER? default: FALSE}

\item{...}{Additional arguments}
}
\value{
Summary printed to console
}
\description{
Summary table for CEEF
}
\keyword{internal}
\keyword{print}
