% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceef_plot_base.R
\name{ceef_plot_base}
\alias{ceef_plot_base}
\title{ceef_plot_base}
\usage{
ceef_plot_base(
  pos,
  scatter.data,
  flip,
  relative,
  dominance,
  ceef.points,
  colour,
  orig.avg
)
}
\description{
ceef_plot_base
}
