% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceef_plot_ggplot.R
\name{ceef_plot_ggplot}
\alias{ceef_plot_ggplot}
\title{ceef_plot_ggplot}
\usage{
ceef_plot_ggplot(
  he,
  ceef.points,
  scatter.data,
  exArgs,
  dominance,
  relative,
  orig.avg,
  colour,
  flip,
  pos,
  ...
)
}
\description{
ceef_plot_ggplot
}
