% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceef_plot_base.R, R/ceef_plot_ggplot.R
\name{ceef_plot_base}
\alias{ceef_plot_base}
\alias{ceef_plot_graph}
\alias{ceef_plot_ggplot}
\title{CEEF plot base R version}
\usage{
ceef_plot_base(he, frontier_data, frontier_params)

ceef_plot_ggplot(he, frontier_data, frontier_params, ...)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{frontier_data}{Frontier data}

\item{frontier_params}{Frontier parameters}

\item{...}{Additional arguments}
}
\description{
CEEF plot base R version

Cost-effectiveness efficiency frontier plot

Choice of base R, ggplot2.

CEEF plot ggplot2 version
}
