% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceplane.plot.R
\name{ceplane.plot}
\alias{ceplane.plot}
\title{Cost-effectiveness Plane Plot}
\usage{
ceplane.plot(he, ...)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{...}{Additional arguments}
}
\description{
Produces a scatter plot of the cost-effectiveness plane,
together with the sustainability area, as a function of
the selected willingness to pay threshold.
}
