% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceplane_geom_params.R
\name{ceplane_geom_params}
\alias{ceplane_geom_params}
\title{Extract Separate Parameter Sets}
\usage{
ceplane_geom_params(...)
}
\arguments{
\item{...}{Additional arguments}
}
\description{
Extract Separate Parameter Sets
}
\keyword{internal}
