% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceplane_ggplot_params.R
\name{ceplane_ggplot_params}
\alias{ceplane_ggplot_params}
\title{CE-plane ggplot Parameters}
\usage{
ceplane_ggplot_params(he, pos_legend, graph_params, ...)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{pos_legend}{Position of legend}

\item{graph_params}{Other graphical parameters}

\item{...}{Additional arguments}
}
\description{
CE-plane ggplot Parameters
}
\keyword{internal}
