% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceplane_plot_ggplot.R
\name{ceplane_plot_ggplot.bcea}
\alias{ceplane_plot_ggplot.bcea}
\title{ceplane_plot_ggplot}
\usage{
\method{ceplane_plot_ggplot}{bcea}(he, wtp, pos_legend, graph_params, ...)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{wtp}{}

\item{pos_legend}{}

\item{graph_params}{}

\item{...}{}
}
\description{
ceplane_plot_ggplot
}
\examples{

data(Vaccine)
he <- bcea(e, c)

ceplane.plot(he, graph = "ggplot2")

data(Smoking)
he <- bcea(e, c, ref = 4, Kmax = 500, interventions = treats)

ceplane.plot(he, graph = "ggplot2")

ceplane.plot(he,
             wtp = 200,
             pos = "right",
             ICER_size = 2,
             graph = "ggplot2")
   
ceplane.plot(he,
             wtp = 200,
             pos = TRUE,
             graph = "ggplot2")

ceplane.plot(he,
             graph = "ggplot2",
             wtp=200,
             theme = theme_linedraw())
             
}
\keyword{hplot}
