% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEriskav.plot.R
\name{CEriskav.plot.CEriskav}
\alias{CEriskav.plot.CEriskav}
\alias{CEriskav.plot}
\title{Cost-effectiveness Plot Including a Parameter of Risk Aversion (CEriskav)}
\usage{
\method{CEriskav.plot}{CEriskav}(
  he,
  pos = "topright",
  graph = c("base", "ggplot2", "plotly"),
  ...
)

CEriskav.plot(he, ...)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{pos}{Legend position}

\item{graph}{A string used to select the graphical engine to use for
plotting. Should (partial-)match the options \code{"base"},
\code{"ggplot2"} or \code{"plotly"}. Default value is \code{"base"}.}

\item{pos_legend}{Legend position}
}
\value{
\item{plot}{ A ggplot or plot_ly object containing the plot. Returned only
if \code{graph_type="ggplot2"} or \code{graph_type="plotly"}.}
}
\description{
Cost-effectiveness Plot Including a Parameter of Risk Aversion (CEriskav)
}
\references{
\insertRef{Baio2013}{BCEA}
}
\seealso{
\code{\link[=bcea]{bcea()}}
}
