% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_xxx.R
\name{comp_names_from_}
\alias{comp_names_from_}
\title{Comparison Names From}
\usage{
comp_names_from_(df_ce)
}
\arguments{
\item{df_ce}{Cost-effectiveness dataframe}
}
\description{
Comparison Names From
}
\keyword{internal}
