% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_xxx.R
\name{compute_CEAC}
\alias{compute_CEAC}
\title{Compute Cost-Effectiveness Acceptability Curve}
\usage{
compute_CEAC(ib)
}
\arguments{
\item{ib}{Incremental benefit}
}
\value{
Array with dimensions (interv x k)
}
\description{
Compute Cost-Effectiveness Acceptability Curve
}
\seealso{
\code{\link[=ceac.plot]{ceac.plot()}}
}
