% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_eib_cri.R
\name{compute_eib_cri}
\alias{compute_eib_cri}
\title{Calculate Credible Intervals}
\usage{
compute_eib_cri(he, alpha_cri = 0.05, cri.quantile = TRUE)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{alpha_cri}{Significance level, 0 - 1}

\item{cri.quantile}{Credible interval quantile?; logical}
}
\value{
cri
}
\description{
For expected incremental benefit plot.
}
