% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_xxx.R
\name{compute_EVI}
\alias{compute_EVI}
\title{Compute Expected Value of Information}
\usage{
compute_EVI(ol)
}
\arguments{
\item{ol}{Opportunity loss}
}
\value{
EVI
}
\description{
Compute Expected Value of Information
}
