% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evppi_helpers.R
\name{compute_evppi}
\alias{compute_evppi}
\title{Compute EVPPI}
\usage{
compute_evppi(he, fit.full)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{fit.full}{fit.full}
}
\value{
list
}
\description{
Compute EVPPI
}
\seealso{
\code{\link{evppi}}
}
\keyword{internal}
