% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_IB.R
\name{compute_IB}
\alias{compute_IB}
\title{Compute Incremental Benefit}
\usage{
compute_IB(df_ce, k)
}
\arguments{
\item{df_ce}{Dataframe of cost and effectiveness deltas}

\item{k}{Vector of willingness to pay values}
}
\value{

}
\description{
Sample of incremental net monetary benefit for each
willingness-to-pay threshold and comparator.
}
\examples{

}
