% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_xxx.R
\name{compute_ICER}
\alias{compute_ICER}
\title{Compute Incremental Cost-Effectiveness Ratio}
\usage{
compute_ICER(df_ce)
}
\arguments{
\item{df_ce}{Cost-effectiveness dataframe}
}
\value{
ICER for all comparisons
}
\description{
Defined as
}
\details{
\deqn{ICER = \Delta_c/\Delta_e}
}
