% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_xxx.R
\name{compute_kstar}
\alias{compute_kstar}
\title{Compute kstar}
\usage{
compute_kstar(k, best, ref)
}
\arguments{
\item{k}{}

\item{best}{}

\item{ref}{}
}
\value{
kstar
}
\description{
Find k when optimal decision changes.
}
