% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_xxx.R
\name{compute_ol}
\alias{compute_ol}
\title{Compute ol}
\usage{
compute_ol(n_sim, K, Ustar, U, best)
}
\arguments{
\item{n_sim}{}

\item{K}{}

\item{Ustar}{}

\item{U}{}

\item{best}{}
}
\value{
ol
}
\description{
Compute ol
}
