% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_xxx.R
\name{compute_p_best_interv}
\alias{compute_p_best_interv}
\title{Compute Probability Best Intervention}
\usage{
compute_p_best_interv(he)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}
}
\description{
Compute Probability Best Intervention
}
