% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_xxx.R
\name{compute_p_optimal_best}
\alias{compute_p_optimal_best}
\title{Compute Probability Optimal Intervention Best}
\usage{
compute_p_optimal_best(he)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}
}
\description{
Compute Probability Optimal Intervention Best
}
