% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_xxx.R
\name{compute_U}
\alias{compute_U}
\title{Compute U Statistic}
\usage{
compute_U(df_ce, k)
}
\arguments{
\item{df_ce}{Cost-effectiveness dataframe}

\item{k}{Willingness to pay vector}
}
\value{
Array with dimensions (sim x k x ints)
}
\description{
Sample of net (monetary) benefit for each
willingness-to-pay threshold and intervention.
}
