% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_xxx.R
\name{compute_Ubar}
\alias{compute_Ubar}
\title{Compute NB for mixture of interventions}
\usage{
compute_Ubar(he, value)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{value}{Mixture weights}
}
\description{
Compute NB for mixture of interventions
}
