% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_xxx.R
\name{compute_Ustar}
\alias{compute_Ustar}
\title{Compute Ustar Statistic}
\usage{
compute_Ustar(U)
}
\arguments{
\item{U}{Net monetary benefit (sim x k x intervs)}

\item{n_sim}{Number of simulations}

\item{K}{Willingness-to-pay length}
}
\value{
Ustar (sim x k)
}
\description{
The maximum utility value among the comparators, indicating which
intervention produced the most benefits at each simulation.
}
