% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_xxx.R
\name{compute_vi}
\alias{compute_vi}
\title{Compute Value of Information}
\usage{
compute_vi(Ustar, U)
}
\arguments{
\item{Ustar}{Maximum utility value (sim x k)}

\item{U}{Net monetary benefit (sim x k x interv)}
}
\value{
Array with dimensions (sim x k)
}
\description{
The difference between the maximum utility computed for the current
parameter configuration \eqn{U^*} and the utility of the intervention which
is associated with the maximum utility overall.
}
\details{
In mathematical notation,
\deqn{VoI(\mathbf{\theta}) = U^*(\mathbf{\theta}) - \mathrm{E}[U(\mathbf{\theta}^{\tau})]}

For each row of the simulations table, the VoI is computed as the
difference between the current value of \eqn{U^*} and the
mean of the entire vector of \eqn{U}.
}
