% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour_ggplot_params.R
\name{contour_ggplot_params}
\alias{contour_ggplot_params}
\title{Contour ggplot Parameters}
\usage{
contour_ggplot_params(he, graph_params, ...)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{graph_params}{Other graphical parameters}

\item{...}{Additional arguments}

\item{pos_legend}{Position of legend}
}
\description{
Contour ggplot Parameters
}
\keyword{internal}
