% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour_base.R, R/contour_ggplot.R
\name{contour_graph}
\alias{contour_graph}
\alias{contour_base}
\alias{contour_ggplot}
\title{Contour Cost-Effectiveness Plane}
\usage{
contour_base(
  he,
  params,
  scale,
  nlevels,
  levels,
  xlim,
  ylim,
  extra_args,
  comparison = 1
)

contour_ggplot(
  he,
  params,
  scale,
  nlevels,
  levels,
  xlim,
  ylim,
  extra_args,
  comparison = 1
)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{params}{Plot parameters}

\item{scale}{Scale}

\item{nlevels}{Number of levels}

\item{levels}{levels}

\item{xlim}{x-axis limits}

\item{ylim}{y-axis limits}

\item{extra_args}{additional arguments}

\item{comparison}{Comparison interventions; default 1}
}
\description{
Choice of base R, ggplot2
}
