% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_pts_to_mm.R
\name{convert_pts_to_mm}
\alias{convert_pts_to_mm}
\title{Use from Base  R to ggplot}
\usage{
convert_pts_to_mm(x)
}
\arguments{
\item{x}{points}
}
\description{
Use from Base  R to ggplot
}
\keyword{internal}
