% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eib.plot.R
\name{eib.plot}
\alias{eib.plot}
\title{Expected Incremental Benefit (EIB) Plot}
\usage{
eib.plot(he, ...)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{...}{If \code{graph="ggplot2"} and a named theme object is supplied,
 it will be added to the ggplot object. Additional arguments:
\itemize{
 \item \code{alpha} can be used to set the CrI level when \code{plot.cri=TRUE},
 with a default value of \code{alpha=0.05}.
 \item \code{cri.quantile} controls the the method of calculation of the credible
 intervals. The default value \code{cri.quantile=TRUE} defines the CrI as the
 interval between the \code{alpha/2}-th and \code{1-alpha/2}-th quantiles of
 the IB distribution. Setting \code{cri.quantile=FALSE} will use a normal
 approximation on the IB distribution to calculate the intervals.
 \item \code{line_colors}: specifies the line colour(s) - all graph types.
 \item \code{line_types}: specifies the line type(s) as lty numeric values - all graph types.
 \item \code{area_include}: include area under the EIB curve - plotly only.
 \item \code{area_color}: specifies the AUC curve - plotly only.}}
}
\value{
\item{eib}{ If \code{graph="ggplot2"} a ggplot object, or if \code{graph="plotly"} 
a plotly object containing the requested plot. Nothing is returned when \code{graph="base"}, 
the default.} The function produces a plot of the
Expected Incremental Benefit as a function of the discrete grid
approximation of the willingness to pay parameter. The break even point
(i.e. the point in which the EIB = 0, i.e. when the optimal decision changes
from one intervention to another) is also showed by default. The value `k*` is
the discrete grid approximation of the ICER.
}
\description{
Produces a plot of the Expected Incremental Benefit (EIB) as a function of
the willingness to pay.
}
\examples{
data(Vaccine)
 
# Runs the health economic evaluation using BCEA
m <- bcea(
      e=e,
      c=c,                  # defines the variables of 
                            #  effectiveness and cost
      ref=2,                # selects the 2nd row of (e, c) 
                            #  as containing the reference intervention
      interventions=treats, # defines the labels to be associated 
                            #  with each intervention
      Kmax=50000,           # maximum value possible for the willingness 
                            #  to pay threshold; implies that k is chosen 
                            #  in a grid from the interval (0, Kmax)
      plot=FALSE             # plots the results
)
eib.plot(m)
eib.plot(m, graph = "ggplot2") + ggplot2::theme_linedraw()

data(Smoking)
treats <- c("No intervention", "Self-help",
            "Individual counselling", "Group counselling")
m <- bcea(e, c, ref = 4, interventions = treats, Kmax = 500)
eib.plot(m)

}
\references{
Baio, G., Dawid, A. P. (2011). Probabilistic Sensitivity
Analysis in Health Economics. Statistical Methods in Medical Research
doi:10.1177/0962280211419832.

Baio G. (2012). Bayesian Methods in Health Economics. CRC/Chapman Hall, London.
}
\seealso{
\code{\link{bcea}},
         \code{\link{ib.plot}},
         \code{\link{ceplane.plot}}
}
\author{
Gianluca Baio, Andrea Berardi
}
\keyword{"Expected}
\keyword{"Health}
\keyword{Benefit"}
\keyword{Incremental}
\keyword{economic}
\keyword{evaluation"}
