% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eib_plot_graph.R
\name{eib_plot_graph}
\alias{eib_plot_graph}
\alias{eib_plot_base}
\alias{eib_plot_ggplot}
\alias{eib_plot_plotly}
\title{Expected Incremental Benefit Plot By Graph Device}
\usage{
eib_plot_base(he, graph_params, ...)

eib_plot_ggplot(he, graph_params, ...)

eib_plot_plotly(he, graph_params, ...)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{graph_params}{Graph parameters}

\item{...}{Additional parameters}
}
\description{
Choice of base R, ggplot2 or plotly.
}
