% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eib_plot_plotly.R
\name{eib_plot_plotly}
\alias{eib_plot_plotly}
\title{eib_plot_plotly}
\usage{
eib_plot_plotly(he, wtp, pos_legend, graph_params)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{wtp}{}

\item{pos_legend}{}

\item{graph_params}{}
}
\description{
eib_plot_plotly
}
