% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evppi_helpers.R
\name{estimate.hyperparams}
\alias{estimate.hyperparams}
\title{Estimate hyperparameters}
\usage{
estimate.hyperparams(x, input.matrix, parameter, n.sim)
}
\arguments{
\item{x}{x}

\item{input.matrix}{Input matrix}

\item{parameter}{Parameters}

\item{n.sim}{Number of simulations}
}
\description{
Estimate hyperparameters
}
\seealso{
\code{\link{evppi}}
}
