% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evi.plot.R
\name{evi.plot}
\alias{evi.plot}
\title{Expected Value of Information (EVI) Plot}
\usage{
evi.plot(he, ...)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{...}{Additional graphical arguments:
\itemize{
  \item \code{line_colors} to specify the EVPI line colour - all graph types.
  \item \code{line_types} to specify the line type (lty) - all graph types.
  \item \code{area_include} to specify whether to include the area under the
  EVPI curve - plotly only.
  \item \code{area_color} to specify the area under the colour curve - plotly only.}}
}
\description{
Plots the Expected Value of Information (EVI) against the willingness to pay.
}
