% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evi_plot_ggplot.R
\name{evi_plot_ggplot}
\alias{evi_plot_ggplot}
\title{EVI plot ggplot version}
\usage{
evi_plot_ggplot(he, data.psa, plot_aes, plot_annotations)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{data.psa}{Data}

\item{plot_aes}{Aesthetic parameters}

\item{plot_annotations}{Plot parameters}
}
\description{
EVI plot ggplot version
}
