% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evi_plot_graph.R
\name{evi_plot_graph}
\alias{evi_plot_graph}
\alias{evi_plot_base}
\alias{evi_plot_ggplot}
\alias{evi_plot_plotly}
\title{Expected Value of Information Plot By Graph Device}
\usage{
evi_plot_base(he, data.psa, plot_aes, plot_annotations)

evi_plot_ggplot(he, data.psa, plot_aes, plot_annotations)

evi_plot_plotly(data.psa, plot_aes, plot_annotations)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{data.psa}{Data}

\item{plot_aes}{Aesthetic parameters}

\item{plot_annotations}{Plot parameters}
}
\description{
Choice of base R, \pkg{ggplot2} or \pkg{plotly}.
}
