% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evi_plot_plotly.R
\name{evi_plot_plotly}
\alias{evi_plot_plotly}
\title{EVI plot plotly version}
\usage{
evi_plot_plotly(data.psa, plot_aes, plot_annotations)
}
\arguments{
\item{data.psa}{Data}

\item{plot_aes}{Aesthetic parameters}

\item{plot_annotations}{Plot parameters}
}
\description{
EVI plot plotly version
}
