% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evppi.default.R
\name{evppi.default}
\alias{evppi.default}
\title{evppi.default}
\usage{
\method{evppi}{default}(param_idx, input, he, N = NULL, plot = FALSE, residuals = TRUE, ...)
}
\arguments{
\item{param_idx}{}

\item{input}{}

\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{N}{}

\item{plot}{}

\item{residuals}{}

\item{...}{}
}
\value{

}
\description{
This function has been completely changed and restructured to make it possible to change regression method.
The method argument can now be given as a list. The first element element in the list is a vector giving the
regression method for the effects. The second gives the regression method for the costs. The `method' argument
can also be given as before which then uses the same regression method for all curves.
All other extra_args can be given as before. 'int.ord' can be updated using the list formulation above to give
the interactions for each different curve.
The formula argument for GAM can only be given once, either 'te()' or 's() + s()' as this is
for computational reasons rather than to aid fit.
You can still plot the INLA mesh elements but not output the meshes.
}
\examples{

}
