% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evppi_plot_base.R, R/evppi_plot_ggplot.R,
%   R/plot.evppi.R
\name{evppi_plot_base.evppi}
\alias{evppi_plot_base.evppi}
\alias{evppi_plot_base}
\alias{evppi_plot_ggplot.evppi}
\alias{evppi_plot_ggplot}
\alias{evppi_plot_graph}
\title{Plot Expected Value of Partial Information With Respect to a
Set of Parameters}
\usage{
\method{evppi_plot_base}{evppi}(evppi_obj, pos_legend = pos, col = NULL, annot = FALSE)

evppi_plot_base(evppi_obj, ...)

\method{evppi_plot_ggplot}{evppi}(evppi_obj, pos_legend = c(0, 0.8), col = c(1, 1), ...)

evppi_plot_ggplot(evppi_obj, ...)
}
\arguments{
\item{evppi_obj}{Object of class evppi}

\item{pos_legend}{Legend position}

\item{col}{Line colours; vector}

\item{annot}{Annotate EVPPI curve with parameter names}

\item{x}{Result of call to \code{createInputs()}}
}
\description{
Plot Expected Value of Partial Information With Respect to a
Set of Parameters
}
