% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evppi_helpers.R
\name{fit.gam}
\alias{fit.gam}
\title{Gaussian Additive Model Fitting}
\usage{
fit.gam(parameter, inputs, x, form)
}
\arguments{
\item{parameter}{Parameter}

\item{inputs}{Inputs}

\item{x}{Response variable}

\item{form}{Formula}
}
\value{
List
}
\description{
Gaussian Additive Model Fitting
}
\seealso{
\code{\link{evppi}}
}
\keyword{internal}
