% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evppi_helpers.R
\name{fit.gp}
\alias{fit.gp}
\title{Fit Gaussian Process}
\usage{
fit.gp(parameter, inputs, x, n.sim)
}
\arguments{
\item{parameter}{Parameters}

\item{inputs}{Inputs}

\item{x}{Response variable}

\item{n.sim}{Number of simulations}
}
\value{
list
}
\description{
Fit Gaussian Process
}
\seealso{
\code{\link{evppi}}
}
