% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_cri.R
\name{geom_cri}
\alias{geom_cri}
\title{Credible interval ggplot geom}
\usage{
geom_cri(plot.cri = TRUE, params = NA)
}
\arguments{
\item{plot.cri}{Should we plot CrI? Logical}

\item{params}{Plot parameters including data}
}
\description{
Credible interval ggplot geom
}
\keyword{aplot}
\keyword{internal}
