% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_contour_quadrants.R
\name{geom_quad_txt}
\alias{geom_quad_txt}
\title{Geom Quadrant Text}
\usage{
geom_quad_txt(he, graph_params)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{graph_params}{Plot parameters; list}
}
\description{
Geom Quadrant Text
}
\keyword{aplot}
\keyword{internal}
