% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evppi.default.R
\name{get_fitted_}
\alias{get_fitted_}
\title{Get fitted values from evppi object}
\usage{
get_fitted_(val, voi_methods, voi_models)
}
\value{
matrix
}
\description{
Get fitted values from evppi object
}
\keyword{internal}
