% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrassmannOptim.R
\name{GrassmannOptim}
\alias{GrassmannOptim}
\title{GrassmannOptim}
\usage{
GrassmannOptim(
  objfun,
  W,
  sim_anneal = FALSE,
  temp_init = 20,
  cooling_rate = 2,
  max_iter_sa = 100,
  eps_conv = 1e-05,
  max_iter = 100,
  eps_grad = 1e-05,
  eps_f = .Machine$double.eps,
  verbose = FALSE
)
}
\arguments{
\item{objfun}{objfun}

\item{W}{W}

\item{sim_anneal}{sim_anneal}

\item{temp_init}{temp_init}

\item{cooling_rate}{cooling_rate}

\item{max_iter_sa}{max_iter_sa}

\item{eps_conv}{eps_conv}

\item{max_iter}{max_iter}

\item{eps_grad}{eps_grad}

\item{eps_f}{eps_f}

\item{verbose}{verbose}
}
\value{
List
}
\description{
This function is taken from the GrassmannOptim package
by Kofi Placid Adragni and Seongho Wu
https://cran.r-project.org/web/packages/GrassmannOptim/index.html
}
\keyword{internal}
