% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceac_boot.R
\name{ib_boot}
\alias{ib_boot}
\title{IB Bootstrapping}
\usage{
ib_boot(he, k, R)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{k}{Willingness to pay}

\item{R}{}
}
\description{
IB Bootstrapping
}
