% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ib_plot_base.R, R/ib_plot_ggplot.R
\name{ib_plot_base}
\alias{ib_plot_base}
\alias{ib_plot_graph}
\alias{ib_plot_ggplot}
\alias{ib_plot_plotly}
\title{IB plot base R version}
\usage{
ib_plot_base(he, comparison, wtp, bw, n, xlim)

ib_plot_ggplot(he, comparison, wtp, bw, n, xlim)

ib_plot_plotly(he, comparison, wtp, bw, n, xlim)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{comparison}{Comparison intervention}

\item{wtp}{Willingness to pay}

\item{bw}{band width}

\item{n}{Number}

\item{xlim}{x-axis limits}
}
\description{
Choice of base R, ggplot2, plotly
}
