% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.rank.R
\name{info.rank.bcea}
\alias{info.rank.bcea}
\title{Information-Rank Plot for bcea Class}
\usage{
\method{info.rank}{bcea}(
  he,
  parameter,
  input,
  wtp = he$k[min(which(he$k >= he$ICER))],
  howManyPars = NULL,
  graph = c("base", "plotly"),
  ...
)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{parameter}{A vector of parameters for which the individual EVPPI
should be calculated. This can be given as a string (or vector of strings)
of names or a numeric vector, corresponding to the column numbers of
important parameters.}

\item{input}{A matrix containing the simulations for all the parameters
monitored by the call to JAGS or BUGS. The matrix should have column names
matching the names of the parameters and the values in the vector parameter
should match at least one of those values.}

\item{wtp}{A value of the wtp for which the analysis should be performed. If
not specified then the break-even point for the current model will be used.}

\item{howManyPars}{Optional maximum number of parameters to be included in the bar plot. 
Includes all parameters by default.}

\item{graph}{A string used to select the graphical enging to use for plotting.
Should (partial-)match one of the two options "base" or "plotly". Default value is "base"}

\item{...}{Extra arguments}
}
\description{
Information-Rank Plot for bcea Class
}
\references{
Baio, G., Dawid, A. P. (2011). Probabilistic Sensitivity
Analysis in Health Economics. Statistical Methods in Medical Research
doi:10.1177/0962280211419832.

Baio G. (2012). Bayesian Methods in Health Economics. CRC/Chapman Hall, London.
}
\seealso{
\code{\link{bcea}},
         \code{\link{evppi}}
}
\author{
Anna Heath, Gianluca Baio, Andrea Berardi
}
\keyword{dplot}
\keyword{models}
