% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_rank_graph.R
\name{info_rank_graph}
\alias{info_rank_graph}
\alias{info_rank_base}
\alias{info_rank_ggplot}
\alias{info_rank_plotly}
\title{Info Rank Plot By Graph Device}
\usage{
info_rank_base(he, params)

info_rank_ggplot(he, params)

info_rank_plotly(params)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{params}{Graph Parameters including data}
}
\description{
Choice of base R, \pkg{ggplot2} and \pkg{plotly}.
}
