% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inforank_params.R
\name{inforank_params}
\alias{inforank_params}
\title{Prepare Info Rank plot parameters}
\usage{
inforank_params(he, inp, wtp = NULL, rel, howManyPars, extra_args)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{inp}{Inputs}

\item{wtp}{Willingness to pay}

\item{rel}{Relative size}

\item{howManyPars}{How mnay parameters to use?}

\item{extra_args}{Additional arguments}
}
\description{
Prepare Info Rank plot parameters
}
\keyword{internal}
