% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helpers.R
\name{is.rel}
\alias{is.rel}
\title{Reports whether x is a rel object
Copied from ggplot2}
\usage{
is.rel(x)
}
\arguments{
\item{x}{An object to test}
}
\description{
Reports whether x is a rel object
Copied from ggplot2
}
\keyword{internal}
