% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kstar_vlines.R
\name{kstar_vlines}
\alias{kstar_vlines}
\title{Prepare K-star vertical lines}
\usage{
kstar_vlines(he, plot_params)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{plot_params}{Plots parameters}
}
\description{
Prepare K-star vertical lines
}
\keyword{internal}
