% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCEAweb.R
\name{launch}
\alias{launch}
\title{Internal launch function}
\usage{
launch(e, c, parameters, ...)
}
\arguments{
\item{e}{effects}

\item{c}{costs}

\item{parameters}{}

\item{...}{}
}
\description{
Internal launch function
}
