% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_labels.R
\name{line_labels}
\alias{line_labels}
\alias{line_labels.default}
\alias{line_labels.pairwise}
\title{Create Labels for Plot}
\usage{
line_labels(he, ...)

\method{line_labels}{default}(he, ref_first = TRUE, ...)

\method{line_labels}{pairwise}(he, ...)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{...}{Additional arguments}
}
\description{
Create Labels for Plot

Swapped labels so that reference is second
}
\keyword{internal}
