% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.barplot.R
\name{make.barplot}
\alias{make.barplot}
\alias{make.barplot_base}
\alias{make.barplot_plotly}
\title{Make Bar Plot for Different Graphic Devices}
\usage{
make.barplot_base(
  scores,
  chk2,
  tit,
  xlab,
  xlim,
  ca,
  cn,
  mai,
  space,
  howManyPars
)

make.barplot_plotly(
  scores,
  chk2,
  tit,
  xlab,
  xlim,
  ca,
  cn,
  mai,
  space,
  howManyPars
)
}
\arguments{
\item{scores}{Scores}

\item{chk2}{chk2}

\item{tit}{tit}

\item{xlab}{x-axis label}

\item{xlim}{x-axis limits}

\item{ca}{ca}

\item{cn}{cn}

\item{mai}{Margins}

\item{space}{space}

\item{howManyPars}{How many parameters}
}
\value{
Bar plot
}
\description{
Base R or plotly.
}
\seealso{
\code{\link{info.rank}}
}
