% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evppi_helpers.R
\name{make.mesh}
\alias{make.mesh}
\title{Make Mesh}
\usage{
make.mesh(data, convex.inner, convex.outer, cutoff, max.edge)
}
\arguments{
\item{data}{Data}

\item{convex.inner}{convex.inner}

\item{convex.outer}{convex.outer}

\item{cutoff}{Cut-off}

\item{max.edge}{Maximum edge}
}
\description{
Make Mesh
}
