% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.report.R
\name{make.report}
\alias{make.report}
\title{make.report}
\usage{
make.report(he, evppi = NULL, ext = "pdf", ...)
}
\arguments{
\item{he}{An object obtained as output to a call to \code{bcea}}

\item{evppi}{An object obtained as output to a call to \code{evppi}
(default is NULL, so not essential to producing the report).}

\item{ext}{A string of text to indicate the extension of the 
resulting output file. Possible options are \code{"pdf"}, \code{"docx"}.
This requires the use of pandoc, knitr and rmarkdown.}

\item{...}{Additional parameters.}
}
\description{
Constructs the automated report from the output of the BCEA
}
\examples{
\donttest{
data(Vaccine)
m=bcea(e,c,ref=2)
makeReport(m)
}
}
\references{
Baio, G., Dawid, A. P. (2011). Probabilistic Sensitivity
Analysis in Health Economics.  Statistical Methods in Medical Research
doi:10.1177/0962280211419832.

Baio G. (2012). Bayesian Methods in Health Economics. CRC/Chapman Hall,
London
}
\seealso{
\code{\link{bcea}}
}
\author{
Gianluca Baio
}
\keyword{"Expected}
\keyword{"Health}
\keyword{economic}
\keyword{evaluation",}
\keyword{information"}
\keyword{of}
\keyword{value}
