% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_legend_plotly.R
\name{make_legend_plotly}
\alias{make_legend_plotly}
\title{legend positioning not great - must be customized case by case}
\usage{
make_legend_plotly(pos_legend)
}
\description{
legend positioning not great - must be customized case by case
}
