% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bcea.R
\name{multiplot}
\alias{multiplot}
\title{multiplot}
\usage{
multiplot(plotlist = NULL, cols = 1, layout = NULL, extra_args, ...)
}
\arguments{
\item{plotlist}{List of ggplot objects}

\item{cols}{Number of columns}

\item{layout}{Matrix}

\item{extra_args}{List of plotting arguments}

\item{...}{Potential further arguments.}
}
\value{
Plot
}
\description{
Arrange plots in grid. Source from R graphics cookbook.
}
