% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_bcea.R
\name{new_bcea}
\alias{new_bcea}
\title{Constructor for bcea}
\usage{
new_bcea(df_ce, k)
}
\arguments{
\item{df_ce}{Dataframe of all simulation eff and cost}

\item{k}{Vector of willingness to pay values}
}
\value{
List object of class bcea.
}
\description{
Constructor for bcea
}
\seealso{
\code{\link[=bcea]{bcea()}}
}
