% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_lines.R
\name{num_lines}
\alias{num_lines}
\alias{num_lines.pairwise}
\alias{num_lines.bcea}
\alias{num_lines.evppi}
\alias{num_lines.CEriskav}
\alias{num_lines.default}
\title{Get number of lines}
\usage{
num_lines(dat)

\method{num_lines}{pairwise}(dat)

\method{num_lines}{bcea}(dat)

\method{num_lines}{evppi}(dat)

\method{num_lines}{CEriskav}(dat)

\method{num_lines}{default}(dat)
}
\arguments{
\item{dat}{Data}
}
\description{
Get number of lines
}
\keyword{internal}
