% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.report.R
\name{openPDF}
\alias{openPDF}
\title{Automatically open pdf output using default pdf viewer}
\usage{
openPDF(file_name)
}
\arguments{
\item{file_name}{String file names for pdf}
}
\description{
Automatically open pdf output using default pdf viewer
}
\keyword{internal}
