% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evppi_helpers.R
\name{plot.mesh}
\alias{plot.mesh}
\title{Plot Mesh}
\usage{
\method{plot}{mesh}(mesh, data, plot)
}
\arguments{
\item{mesh}{Mesh}

\item{data}{Data}

\item{plot}{Plot; logical}
}
\description{
Plot Mesh
}
