% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eib_cri.R
\name{plot_eib_cri}
\alias{plot_eib_cri}
\title{Plot Credible Intervals}
\usage{
plot_eib_cri(he, params)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{params}{Graph parameters}
}
\description{
Bayesian posterior credible intervals against willingness to pay.
}
\keyword{internal}
