% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evppi_helpers.R
\name{plot_mesh}
\alias{plot_mesh}
\title{Mesh Plot}
\usage{
plot_mesh(mesh, data, plot, ...)
}
\arguments{
\item{mesh}{Mesh}

\item{data}{Data}

\item{plot}{Create plot? logical}

\item{...}{Additional parameters}
}
\description{
Option of interactively saving the plot.
}
\seealso{
\code{\link{evppi}}
}
\keyword{internal}
