% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evppi_helpers.R
\name{prep.x}
\alias{prep.x}
\title{Prepare parameter}
\usage{
prep.x(he, seq_rows, k, l)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{seq_rows}{}

\item{k}{}

\item{l}{}
}
\description{
Prepare parameter
}
