% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_ceplane_params.R
\name{prep_ceplane_params}
\alias{prep_ceplane_params}
\title{prep_ceplane_params}
\usage{
prep_ceplane_params(he, wtp, ...)
}
\arguments{
\item{he}{}

\item{...}{}
}
\value{

}
\description{
In ggplot format, combine user supplied with defaults.
}
\examples{

}
