% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_contour_params.R
\name{prep_contour_params}
\alias{prep_contour_params}
\title{Prepare contour plot parameters
Additional to ceplane parameters}
\usage{
prep_contour_params(he, ...)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{...}{Additional parameters}
}
\value{
A list of parameters
}
\description{
Prepare contour plot parameters
Additional to ceplane parameters
}
\keyword{internal}
