% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_delta_ce.R
\name{prep_delta_ce}
\alias{prep_delta_ce}
\title{Reshape BCEA object in long format (delta_ce)}
\usage{
prep_delta_ce(he)
}
\arguments{
\item{he}{a BCEA object of interest}
}
\value{
a data.frame, displaying the BCEA object information in long format
}
\description{
Reshape BCEA object in long format (delta_ce)
}
\keyword{internal}
