% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_eib_params.R
\name{prep_eib_params}
\alias{prep_eib_params}
\title{Prepare EIB plot parameters}
\usage{
prep_eib_params(he, plot.cri, ...)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{plot.cri}{Make title? Logical}

\item{...}{Additional parameters}
}
\description{
Parameters general to all plotting devices
}
