% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_frontier_data.R
\name{prep_frontier_data}
\alias{prep_frontier_data}
\title{Prepare frontier data}
\usage{
prep_frontier_data(he, threshold = NULL, start.origin = TRUE)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{threshold}{Cost-effectiveness threshold i.e angle of line. Must be >=0 or NULL.}

\item{start.origin}{Where should the frontier start from?}
}
\value{
List with scatter.data, ceef.points, orig.avg
}
\description{
Prepare frontier data
}
\seealso{
ceef.plot
}
\keyword{internal}
