% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evppi_helpers.R
\name{prepare.output}
\alias{prepare.output}
\title{Prepare output}
\usage{
prepare.output(parameters, inputs)
}
\arguments{
\item{parameters}{Parameters}

\item{inputs}{Inputs}
}
\description{
Prepare output
}
\seealso{
\code{\link{evppi}}
}
