% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_methods.R
\name{print.bcea}
\alias{print.bcea}
\title{bcea Print Method}
\usage{
\method{print}{bcea}(x, digits = getOption("digits"), give.attr = FALSE, no.list = TRUE, ...)
}
\arguments{
\item{digits}{Minimal number of significant digits, see \code{\link{print.default}}.}

\item{give.attr}{Logical; if TRUE (default), show attributes as sub structures.}

\item{no.list}{Logical; if TRUE, no ‘list of ...’ nor the class are printed.}

\item{...}{Potential further arguments.}

\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}
}
\description{
bcea Print Method
}
\examples{
data("Vaccine")
he <- BCEA::bcea(e, c)

}
\keyword{print}
