% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrant_params.R
\name{quadrant_params}
\alias{quadrant_params}
\title{Quadrant Parameters
requires just a single comparison group}
\usage{
quadrant_params(he, params)
}
\description{
Quadrant Parameters
requires just a single comparison group
}
\keyword{aplot}
\keyword{internal}
