% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.report.R
\name{quiet}
\alias{quiet}
\title{Allow disabling of the cat messages}
\usage{
quiet(x)
}
\arguments{
\item{x}{Object to quietly return}
}
\description{
Allow disabling of the cat messages
}
\keyword{internal}
