% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEriskav.default.R
\name{reset_bcea}
\alias{reset_bcea}
\title{Resets a modified \code{bcea} object}
\usage{
reset_bcea(he)
}
\arguments{
\item{he}{A modified \code{bcea} object, which now is also in the class \code{mixedAn},
or \code{CEriskav} or both.
returns The original \code{bcea} object, stripped of extra elements (created by
\code{mixedAn} or \code{CEriskav} or both), so that the original methods apply}
}
\description{
Useful to return to a standard \code{bcea} object, after having applied the
setters for \code{mixedAn} or \code{CEriskav}
}
\examples{
data(Vaccine)
# Creates the "standard" BCEA object
m <- bcea(eff, cost, ref=2, interventions=treats)
class(m)

# Applies analysis for risk aversion
CEriskav(m) <- c(0.001, 0.002, 0.003)
# Now m has changed nature
class(m)
# The default method for plot will be the specialised `CEriskav`, but can 
# also specifically request a method, like
# plot.bcea(m)

# More simply, we can revert to the original object
m <- reset_bcea(m)
class(m)

}
\references{
\insertRef{Baio2011}{BCEA}

\insertRef{Baio2013}{BCEA}
}
\seealso{
\code{\link[=bcea]{bcea()}},
\code{\link[=mixedAn]{mixedAn()}},
\code{\link[=CEriskav]{CEriskav()}}
}
\author{
Gianluca Baio
}
