% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setComparisons.R
\name{setComparisons<-}
\alias{setComparisons<-}
\title{Set Comparison Group}
\usage{
setComparisons(he) <- value
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{value}{Comparison}
}
\value{
bcea-type object
}
\description{
One of the alternative way to set (e,c) comparison group.
}
\seealso{
\code{\link{setComparisons}}
}
