% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setComparisons.R
\name{setComparisons}
\alias{setComparisons}
\title{Set Comparisons Group}
\usage{
setComparisons(he, comparison)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{comparison}{Selects the comparator, in case of more than two
interventions being analysed. Default as NULL plots all the comparisons
together. Any subset of the possible comparisons can be selected (e.g.,
\code{comparison=c(1,3)} or \code{comparison=2}).}
}
\description{
One of the alternative way to set (e,c) comparison group.
Simply recompute all comparisons and drop unwanted.
}
\seealso{
\code{\link{setComparisons<-}}
}
