% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcea_setters.R
\name{setKmax<-}
\alias{setKmax<-}
\title{Set Maximum Willingness to Pay}
\usage{
setKmax(he) <- value
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{Kmax}{Maximum willingness to pay}
}
\value{

}
\description{
Set Maximum Willingness to Pay
}
