% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcea_setters.R
\name{setKmax_assign}
\alias{setKmax_assign}
\alias{setKmax<-}
\alias{setKmax<-.bcea}
\alias{setKmax<-.default}
\title{Set Maximum Willingness to Pay

Alternative way to define `K` statistic.}
\usage{
setKmax(he) <- value

\method{setKmax}{bcea}(he) <- value

\method{setKmax}{default}(he) <- value
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{value}{Maximum willingness to pay}
}
\value{
bcea-type object
}
\description{
Set Maximum Willingness to Pay

Alternative way to define `K` statistic.
}
