% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcea_setters.R
\name{setReferenceGroup<-}
\alias{setReferenceGroup<-}
\title{Set Reference Group}
\usage{
setReferenceGroup(he) <- value
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{ref}{Reference group number}
}
\value{

}
\description{
Set Reference Group
}
