% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcea_setters.R
\name{setReferenceGroup_assign}
\alias{setReferenceGroup_assign}
\alias{setReferenceGroup<-}
\alias{setReferenceGroup<-.bcea}
\alias{setReferenceGroup<-.default}
\title{Set Reference Group}
\usage{
setReferenceGroup(he) <- value

\method{setReferenceGroup}{bcea}(he) <- value

\method{setReferenceGroup}{default}(he) <- value
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{value}{Reference group number}
}
\value{
bcea-type object
}
\description{
Alternative way to define (e,c) reference group.
}
