% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{statins_HC}
\alias{statins_HC}
\alias{Statins}
\alias{robust}
\alias{model}
\title{Data set for the Bayesian model for the evidence-synthesis of statins
(robust model using Half Cauchy priors for the structured effects standard
deviations)}
\format{
A BUGS object containing the simulations for the evidence synthesis
model
\describe{
\item{list("statins_HC")}{a BUGS object}
}
}
\source{
A systematic review and economic evaluation of statins for the
prevention of coronary events. Ward 2007.
}
\description{
This data set contains the results of the Bayesian analysis used to model
the effectiveness of various statins. The analysis is based on the simplest
model, using robust priors.
}
\references{
Baio G. (2012). Bayesian Methods in Health Economics.
CRC/Chapman Hall, London
}
\keyword{datasets}
