% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bcea.R
\name{summary.bcea}
\alias{summary.bcea}
\title{Summary Method for Objects of Class \code{bcea}}
\usage{
\method{summary}{bcea}(he, wtp = 25000, ...)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{wtp}{The value of the willingness to pay threshold used in the
summary table.}

\item{...}{Additional arguments affecting the summary produced.}
}
\value{
Prints a summary table with some information on the health economic
output and synthetic information on the economic measures (EIB, CEAC, EVPI).
}
\description{
Produces a table printout with some summary results of the health economic
evaluation.
}
\examples{

data(Vaccine)

summary(he)

}
\references{
Baio, G., Dawid, A. P. (2011). Probabilistic Sensitivity
Analysis in Health Economics. Statistical Methods in Medical Research
doi:10.1177/0962280211419832.

Baio G. (2012). Bayesian Methods in Health Economics. CRC/Chapman Hall, London.
}
\seealso{
\code{\link{bcea}}
}
\author{
Gianluca Baio
}
\keyword{Health}
\keyword{economic}
\keyword{evaluation}
