% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceplane_plot_plotly.R
\name{tabulate_means}
\alias{tabulate_means}
\title{Calculates dataset for ICERs from bcea object}
\usage{
tabulate_means(he, comparisons.label = NULL)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{comparisons.label}{Optional vector of strings with comparison labels}
}
\value{
A data.frame object including mean outcomes, comparison identifier,
  comparison label and associated ICER
}
\description{
Calculates dataset for ICERs from bcea object
}
