% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bcea.R
\name{theme_BCEA}
\alias{theme_BCEA}
\title{A Custom ggplot2 Theme for BCEA}
\usage{
theme_BCEA()
}
\value{
A ggplot2 theme object that can be added to a ggplot.
}
\description{
This theme is designed for use with plots created using the \code{BCEA} package.
It builds on \code{theme_bw()} and customizes axis text, titles,
plot background, and legend styling.
}
\examples{
library(ggplot2)
library(BCEA)
ggplot(mtcars, aes(wt, mpg)) +
 geom_point() +
 theme_BCEA() 
}
