% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes_ggplot.R
\name{theme_bcea}
\alias{theme_bcea}
\alias{theme_default}
\alias{theme_ceac}
\alias{theme_ceplane}
\alias{theme_eib}
\alias{theme_contour}
\title{bcea theme ggplot2}
\usage{
theme_default()

theme_ceac()

theme_ceplane()

theme_eib()

theme_contour()
}
\description{
bcea theme ggplot2
}
\keyword{aplot}
\keyword{internal}
