% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_bcea.R
\name{validate_bcea}
\alias{validate_bcea}
\title{validate_bcea}
\usage{
validate_bcea(eff, cost, ref, interventions)
}
\arguments{
\item{eff}{}

\item{cost}{}

\item{ref}{}

\item{interventions}{}
}
\value{

}
\description{
validate_bcea
}
