% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_eib_params.R
\name{validate_eib_params}
\alias{validate_eib_params}
\title{Validate EIB parameters}
\usage{
validate_eib_params(params)
}
\arguments{
\item{params}{Graph parameters}
}
\value{
List of graph parameters
}
\description{
Validate EIB parameters
}
\seealso{
\code{\link[=prep_eib_params]{prep_eib_params()}}
}
\keyword{internal}
