% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_contour_quadrants.R
\name{add_contour_quadrants}
\alias{add_contour_quadrants}
\title{Add Contour Quadrants}
\usage{
add_contour_quadrants(he, params)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{params}{List}
}
\value{
Plot side effect
}
\description{
Add Contour Quadrants
}
\keyword{aplot}
\keyword{internal}
