\name{BCEA-package}
\alias{BCEA-package}
\alias{BCEA}
\docType{package}
\title{
BCEA: A package for Bayesian Cost-Effectiveness Analysis
}
\description{
A package to post-process the results of a Bayesian health economic model and produce 
standardised output for the analysis of the results
}
\details{
\tabular{ll}{
Package: \tab BCEA\cr
Type: \tab Package\cr
Version: \tab 2.2-6\cr
Date: \tab 2017-10-02\cr
License: \tab GPL2 \cr
LazyLoad: \tab yes\cr
}
Given the results of a Bayesian model (possibly based on MCMC) in the form of 
simulations from the posterior distributions of suitable variables of costs and clinical 
benefits for two or more interventions, produces a health economic evaluation. Compares 
one of the interventions (the "reference") to the others ("comparators"). Produces many 
summary and plots to analyse the results
}
\author{
Gianluca Baio, Andrea Berardi, Anna Heath

Maintainer: Gianluca Baio <gianluca@stats.ucl.ac.uk>
}
\references{
Baio, G., Dawid, A. P. (2011). Probabilistic Sensitivity Analysis in Health Economics. 
Statistical Methods in Medical Research doi:10.1177/0962280211419832.

Baio G. (2012). Bayesian Methods in Health Economics. CRC/Chapman Hall, London

Baio G., Berardi A., Heath A. (forthcoming). Bayesian Cost Effectiveness Analysis with 
the R package BCEA. Springer
}
\keyword{Bayesian models}
\keyword{Health economic evaluation}

