% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceef_plot_graph.R
\name{ceef_plot_graph}
\alias{ceef_plot_graph}
\alias{ceef_plot_ggplot}
\alias{ceef_plot_plotly}
\alias{ceef_plot_base}
\title{Cost-effectiveness Efficiency Frontier Plot By Graph Device}
\usage{
ceef_plot_ggplot(he, frontier_data, frontier_params, ...)

ceef_plot_plotly(he, frontier_data, frontier_params, ...)

ceef_plot_base(he, frontier_data, frontier_params)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{frontier_data}{Frontier data}

\item{frontier_params}{Frontier parameters}

\item{...}{Additional arguments}
}
\description{
Choice of base R, \pkg{ggplot2}.
}
