% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_xxx.R
\name{compute_ceaf}
\alias{compute_ceaf}
\title{Compute Cost-Effectiveness Acceptability Frontier}
\usage{
compute_ceaf(p_best_interv)
}
\arguments{
\item{p_best_interv}{Probability of being best intervention}
}
\description{
Compute Cost-Effectiveness Acceptability Frontier
}
