% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_IB.R
\name{compute_IB}
\alias{compute_IB}
\title{Compute Incremental Benefit}
\usage{
compute_IB(df_ce, k)
}
\arguments{
\item{df_ce}{Dataframe of cost and effectiveness deltas}

\item{k}{Vector of willingness to pay values}
}
\value{
Array with dimensions (k x sim x ints)
}
\description{
Sample of incremental net monetary benefit for each
willingness-to-pay threshold, \eqn{k}, and comparator.
}
\details{
In mathematical notation,

\deqn{INMB = k \Delta_e - \Delta_c}.
}
\examples{

}
