% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_xxx.R
\name{compute_kstar}
\alias{compute_kstar}
\title{Compute kstar}
\usage{
compute_kstar(k, best, ref)
}
\arguments{
\item{k}{Willingness-to-pay vector}

\item{best}{Best intervention for each k}

\item{ref}{Reference intervention}
}
\value{
Array with dimensions
}
\description{
Find willingness-to-pay threshold when optimal decision changes.
}
\seealso{
\code{\link{ceac.plot}}
}
