% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_xxx.R
\name{compute_Ustar}
\alias{compute_Ustar}
\title{Compute Ustar statistic}
\usage{
compute_Ustar(n_sim, K, U)
}
\arguments{
\item{n_sim}{}

\item{K}{}

\item{U}{}
}
\value{
Ustar
}
\description{
Compute Ustar statistic
}
