% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_xxx.R
\name{compute_vi}
\alias{compute_vi}
\title{Compute Value of Information}
\usage{
compute_vi(Ustar, U)
}
\arguments{
\item{Ustar}{Maximum utility value}

\item{U}{Net monetary benefit}

\item{n_sim}{Number of simulations}

\item{K}{Willingness-to-pay length}
}
\value{
vi
}
\description{
The difference between the maximum utility computed for the current
parameter configuration U* and the utility of the intervention which
is associated with the maximum utility overall.
}
\details{
For each row of the simulations table, the
VI is computed as the difference between the current value of \eqn{U^*} and the
mean of the entire vector U2. Negative values of the VI imply that for
those simulation specific parameter values treatment options are less
valuable than the current optimal decision.
}
