% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_xxx.R
\name{compute_vi}
\alias{compute_vi}
\title{Compute Value of Information}
\usage{
compute_vi(n_sim, K, Ustar, U)
}
\arguments{
\item{n_sim}{}

\item{K}{}

\item{Ustar}{}

\item{U}{}
}
\value{
vi
}
\description{
Compute Value of Information
}
