% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour_graph.R
\name{contour_graph}
\alias{contour_graph}
\alias{contour_base}
\alias{contour_ggplot}
\title{Contour Cost-Effectiveness Plane}
\usage{
contour_base(he, pos_legend, graph_params, ...)

contour_ggplot(he, pos_legend, graph_params, ...)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{graph_params}{Plot parameters; list}

\item{...}{Additional arguments}

\item{pos}{Parameter to set the position of the legend (only relevant for
multiple interventions, ie more than 2 interventions being compared).
Can be given in form
of a string \code{(bottom|top)(right|left)} for base graphics and
\code{bottom|top|left|right} for ggplot2. It can be a two-elements vector,
which specifies the relative position on the x and y axis respectively, or
alternatively it can be in form of a logical variable, with \code{FALSE}
indicating to use the default position and \code{TRUE} to place it on the
bottom of the plot.}
}
\description{
Choice of base R, ggplot2
}
\seealso{
\code{\link{contour}}
}
