% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour_graph.R
\name{contour_graph}
\alias{contour_graph}
\alias{contour_base}
\alias{contour_ggplot}
\title{Contour Cost-Effectiveness Plane}
\usage{
contour_base(he, pos_legend, graph_params, ...)

contour_ggplot(he, pos_legend, graph_params, ...)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{pos_legend}{Legend position}

\item{graph_params}{Plot parameters; list}

\item{...}{Additional arguments}

\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{pos_legend}{Legend position}

\item{graph_params}{Plot parameters; list}

\item{...}{Additional arguments}
}
\description{
Choice of base R, ggplot2
}
\seealso{
\code{\link{contour}}
}
