% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour_base.R, R/contour_ggplot.R
\name{contour_graph}
\alias{contour_graph}
\alias{contour_base}
\alias{contour_ggplot}
\title{Contour Cost-Effectiveness Plane}
\usage{
contour_base(he, graph_params, extra_args, comparison = 1)

contour_ggplot(he, params, extra_args, ...)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{graph_params}{Plot parameters}

\item{extra_args}{Additional arguments}

\item{comparison}{Comparison interventions; default 1}

\item{params}{Plot parameters}
}
\description{
Choice of base R, ggplot2
}
