% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour_plotly_lines.R
\name{contour_plotly_lines}
\alias{contour_plotly_lines}
\title{Add contour lines to a plotly contour or CE plane plot}
\usage{
contour_plotly_lines(plotly, he, delta_ce, graph_params, pt_cols)
}
\arguments{
\item{plotly}{plotly object generated by calling function}

\item{he}{bcea object of interest}

\item{delta_ce}{Object generated by calling function (long format)}

\item{graph_params}{Other graphical parameters}

\item{pt_cols}{Vector of colours for the interventions}
}
\description{
Add contour lines to a plotly contour or CE plane plot
}
\seealso{
\code{\link[=contour_plotly]{contour_plotly()}}
}
\keyword{internal}
