% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eib_plot_base.R
\name{eib_plot_base}
\alias{eib_plot_base}
\title{EIB plot base R version}
\usage{
eib_plot_base(he, graph_params, ...)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{graph_params}{List of graph parameters}

\item{...}{}
}
\description{
EIB plot base R version
}
