% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eib_plot_base.R
\name{eib_plot_base}
\alias{eib_plot_base}
\title{eib_plot_base}
\usage{
eib_plot_base(he, graph_params, ...)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}
}
\description{
eib_plot_base
}
