% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eib_plot_plotly.R
\name{eib_plot_plotly}
\alias{eib_plot_plotly}
\title{EIB plot plotly version}
\usage{
eib_plot_plotly(
  he,
  alt.legend,
  plot_aes,
  plot_annotations,
  plot.cri,
  cri.quantile,
  comparison,
  alpha,
  cri,
  size = NULL
)
}
\description{
EIB plot plotly version
}
