% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evppi_plot_base.R, R/evppi_plot_ggplot.R
\name{evppi_plot_graph}
\alias{evppi_plot_graph}
\alias{evppi_plot_base}
\alias{evppi_plot_ggplot}
\title{Plot Expected Value of Partial Information With Respect to a
Set of Parameters}
\usage{
evppi_plot_base(evppi_obj, pos_legend, col = NULL, annot = FALSE)

evppi_plot_ggplot(evppi_obj, pos_legend = c(0, 0.8), col = c(1, 1), ...)
}
\arguments{
\item{evppi_obj}{Object of class evppi}

\item{pos_legend}{Position of legend}

\item{col}{Colour}

\item{annot}{Annotate EVPPI curve with parameter names}

\item{...}{Additional arguments}
}
\description{
Base R and \pkg{ggplot2} versions.
}
