% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.rank.R
\name{info.rank}
\alias{info.rank}
\title{Information-Rank Plot}
\usage{
info.rank(he, ...)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{...}{Additional options. These include graphical parameters that the
user can specify:
\itemize{
\item \code{xlim} = limits of the x-axis; ca = font size for the axis
label (default = 0.7 of full size).
\item \code{cn} = font size for the parameter names
vector (default = 0.7 of full size) - base graphics only.
\item \code{mai} = margins of the graph (default = c(1.36, 1.5, 1,1)) - base graphics only.
\item \code{rel} = logical argument that specifies whether the ratio of
EVPPI to EVPI (\code{rel=TRUE}, default) or the absolute value of the EVPPI
should be used for the analysis.
}}
}
\value{
\item{res}{With base graphics: A data.frame containing the ranking of the parameters
  with the value of the selected summary, for the chosen wtp; with plotly: a plotly object, 
  incorporating in the $rank element the data.frame as above.}
  The function produces a 'Info-rank' plot. This is an extension of standard 'Tornado
  plots' and presents a ranking of the model parameters in terms of their
  impact on the expected value of information. For each parameter, the
  specific individual EVPPI is computed and used to measure the impact of
  uncertainty in that parameter over the decision-making process, in terms of
  how large the expected value of gaining more information is.
}
\description{
Produces a plot similar to a Tornado-plot, but based on the analysis of the
EVPPI. For each parameter and value of the willingness-to-pay threshold, a
barchart is plotted to describe the ratio of EVPPI (specific to that
parameter) to EVPI. This represents the relative `importance' of each
parameter in terms of the expected value of information.
}
