% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_methods.R
\name{is.bcea}
\alias{is.bcea}
\title{Check bcea Class}
\usage{
is.bcea(he)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}
}
\value{
\code{is.bcea} returns TRUE or FALSE depending on whether its argument
is a bcea class object.
}
\description{
Check bcea Class
}
