% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.barplot.R
\name{make.barplot}
\alias{make.barplot}
\alias{make.barplot_base}
\alias{make.barplot_plotly}
\title{Make Bar Plot for Different Graphic Devices}
\usage{
make.barplot_base(
  scores,
  chk2,
  tit,
  xlab,
  xlim,
  ca,
  cn,
  mai,
  space,
  howManyPars
)

make.barplot_plotly(scores, chk2, tit, xlab, xlim, mai, space, howManyPars)
}
\arguments{
\item{scores}{Scores}

\item{chk2}{chk2}

\item{tit}{Main title}

\item{xlab}{x-axis label}

\item{xlim}{x-axis limits}

\item{ca}{The magnification to be used for axis annotation relative to the current setting of cex.}

\item{cn}{Size of axis labels}

\item{mai}{Margins}

\item{space}{space}

\item{howManyPars}{Maximum number of parameters}
}
\value{
Bar plot
}
\description{
Base R or plotly versions.
}
\seealso{
\code{\link{info.rank}}
}
