% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_legend_plotly.R
\name{make_legend_plotly}
\alias{make_legend_plotly}
\title{Legend Positioning}
\usage{
make_legend_plotly(pos_legend)
}
\arguments{
\item{pos_legend}{Position of legend}
}
\value{
String
}
\description{
Legend Positioning
}
\keyword{internal}
