% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Plot Multiple bcea Graphs}
\usage{
multiplot(plotlist = NULL, cols = 1, layout_config = NULL)
}
\arguments{
\item{plotlist}{List of ggplot objects}

\item{cols}{Number of columns}

\item{layout_config}{Matrix of plot configuration}
}
\value{
ggplot TableGrob object
}
\description{
Arrange plots in grid. Sourced from R graphics cookbook.
}
\keyword{internal}
