% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evppi_helpers.R
\name{prep.x}
\alias{prep.x}
\title{Prepare Delta arrays}
\usage{
prep.x(he, seq_rows, k, l)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{seq_rows}{Rows of (e,c) to keep}

\item{k}{e or c? 1 or 2.}

\item{l}{Columns of (e,c) to keep}
}
\description{
Prepare Delta arrays
}
\seealso{
\code{\link{evppi}}
}
\keyword{internal}
