% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_ceplane_params.R
\name{prep_ceplane_params}
\alias{prep_ceplane_params}
\title{Prepare CE-plane Parameters}
\usage{
prep_ceplane_params(he, wtp, ...)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{wtp}{Willingness-to-pay}

\item{...}{Additional arguments}
}
\value{
List pf graph parameters
}
\description{
In ggplot format, combine user-supplied
parameters with defaults.
}
