% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_ceplane_params.R
\name{prep_ceplane_params}
\alias{prep_ceplane_params}
\title{Prepare CE-plane Parameters}
\usage{
prep_ceplane_params(he, wtp_params, ...)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{wtp_params}{Willingness-to-pay parameters. This can be a single value or a list.}

\item{...}{Additional arguments}
}
\value{
List of graph parameters
}
\description{
In ggplot format, combine user-supplied
parameters with defaults.
}
\keyword{internal}
