% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_eib_plot_params.R
\name{prep_eib_plot_params}
\alias{prep_eib_plot_params}
\title{prep_eib_plot_params}
\usage{
prep_eib_plot_params(he, ...)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{...}{}
}
\description{
prep_eib_plot_params
}
\keyword{dplot}
