% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_methods.R
\name{print.bcea}
\alias{print.bcea}
\title{data("Vaccine")
he <- BCEA::bcea(e, c)}
\usage{
\method{print}{bcea}(x, digits = getOption("digits"), give.attr = FALSE, no.list = TRUE, ...)
}
\description{
data("Vaccine")
he <- BCEA::bcea(e, c)
}
