% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcea_setters.R
\name{setReferenceGroup<-}
\alias{setReferenceGroup<-}
\title{setReferenceGroup}
\usage{
setReferenceGroup(he, ref) <- value
}
\arguments{
\item{he}{}

\item{ref}{}
}
\value{

}
\description{
setReferenceGroup
}
