% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcea_setters.R
\name{setReferenceGroup<-}
\alias{setReferenceGroup<-}
\title{Set Reference Group}
\usage{
setReferenceGroup(he) <- value
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{value}{Reference group number}
}
\value{
bcea-type objects
}
\description{
Alternative way to define (e,c) reference group.
}
