% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{statins_HC}
\alias{statins_HC}
\title{Statins Evidence-Synthesis Data (Robust Model)}
\format{
A \code{BUGS} object containing the simulations for the evidence synthesis model.
}
\source{
A systematic review and economic evaluation of statins for the
prevention of coronary events. Ward 2007.
}
\description{
This data set contains the results of a Bayesian analysis used to model
the effectiveness of various statins. The analysis uses robust Half-Cauchy
priors for the structured effects standard deviations.
}
\examples{
data(statins_HC)
lapply(statins_HC$sims.list, summary)
}
\references{
Baio G. (2012). \emph{Bayesian Methods in Health Economics}. CRC/Chapman & Hall, London.
}
\keyword{datasets}
