% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct.psa.R
\name{struct.psa}
\alias{struct.psa}
\title{Structural Probability Sensitivity Analysis}
\usage{
struct.psa(
  models,
  effect,
  cost,
  ref = NULL,
  interventions = NULL,
  Kmax = 50000,
  plot = FALSE,
  w = NULL
)
}
\arguments{
\item{models}{A (possibly named) list containing the output from either
R2jags or R2WinBUGS for all the models that need to be combined in the model
average}

\item{effect}{A list containing the measure of effectiveness computed from
the various models (one matrix with \code{n_sim} x \code{n_ints} simulations for each
model)}

\item{cost}{A list containing the measure of costs computed from the various
models (one matrix with \code{n_sim} x \code{n_ints} simulations for each model)}

\item{ref}{Which intervention is considered to be the reference
strategy. The default value \code{ref=1} means that the intervention
appearing first is the reference and the other(s) is(are) the comparator(s)}

\item{interventions}{Defines the labels to be associated with each
intervention. By default and if \code{NULL}, assigns labels in the form
"Intervention1", ... , "InterventionT"}

\item{Kmax}{Maximum value of the willingness to pay to be considered.
Default value is \code{50000}. The willingness to pay is then approximated
on a discrete grid in the interval \verb{[0, Kmax]}. The grid is equal to
\code{k} if the parameter is given, or composed of \code{501} elements if
\code{k=NULL} (the default)}

\item{plot}{A logical value indicating whether the function should produce
the summary plot or not}

\item{w}{A vector of weights. By default it's \code{NULL} to indicate that the
function will calculate the model weights based on DIC and the individual
model fit. This behaviour can be overridden by passing a vector \code{w},
for instance based on expert opinion}
}
\value{
List object of bcea object, model weights and DIC
}
\description{
Computes the weights to be associated with a set of competing models in
order to perform structural PSA.
}
\details{
The model is a list containing the output from either R2jags or
R2WinBUGS for all the models that need to be combined in the model average
effect is a list containing the measure of effectiveness computed from the
various models (one matrix with \code{n_sim} x \code{n_ints} simulations for each model)
cost is a list containing the measure of costs computed from the various
models (one matrix with \code{n_sim} x \code{n_ints} simulations for each model).
}
\examples{

# load sample jags output
data("statins_base")
data("statins_HC")

interventions <- c("Atorvastatin", "Fluvastatin",
                   "Lovastatin", "Pravastatin",
                   "Rosuvastatin", "Simvastatin")

m1 <- bcea(eff = statins_base$sims.list$effect,
           cost = statins_base$sims.list$cost.tot,
           ref = 1, interventions = interventions)

m2 <- bcea(eff = statins_HC$sims.list$effect,
           cost = statins_HC$sims.list$cost.tot,
           ref = 1, interventions = interventions)

models <- list("Base"=statins_base, "Half Cauchy"=statins_HC)

effects <- list(statins_base$sims.list$effect,
                statins_HC$sims.list$effect)
costs <- list(statins_base$sims.list$cost.tot,
              statins_HC$sims.list$cost.tot)

\dontrun{
m3 <- struct.psa(models, effects, costs,
                 ref = 1, interventions = interventions)
}

}
\references{
\insertRef{Baio2013}{BCEA}
}
\seealso{
\code{\link[=bcea]{bcea()}}
}
\author{
Gianluca Baio
}
