% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceplane_plot_plotly.R
\name{tabulate_means}
\alias{tabulate_means}
\title{Calculate Dataset For ICERs From bcea Object}
\usage{
tabulate_means(he, comp_label = NULL)
}
\arguments{
\item{he}{A \code{bcea} object containing the results of the Bayesian
modelling and the economic evaluation.}

\item{comp_label}{Optional vector of strings with comparison labels}
}
\value{
A data.frame object including mean outcomes, comparison identifier,
  comparison label and associated ICER
}
\description{
Calculate Dataset For ICERs From bcea Object
}
