% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_bcea.R
\name{validate_bcea}
\alias{validate_bcea}
\title{Validate bcea}
\usage{
validate_bcea(eff, cost, ref, interventions)
}
\arguments{
\item{eff}{Effectiveness matrix}

\item{cost}{Cost matrix}

\item{ref}{Reference intervention}

\item{interventions}{All interventions}
}
\description{
Validate bcea
}
\keyword{internal}
